Update predresults.R
I eliminated repetitive code. I used nested functions for readability and to have the main function be more modular for future features and maintenance. Dplyr code was changed to base R to avoid decency and depreciation issues. The pre/post feature is no longer broken. However, the 'mixed' feature still needs to be worked on. I've left the original code block for that feature commented out below the function calls on the example data. - Richard