Update predresults.R
eliminated the 'cols_to_compare' as it wasn't needed. I also noticed a potential problem that has an easy fix once determined what do with it. Do we want the differences of the means/medians or the means/medians of the differences? The two nested functions calculating the difference vector do opposites of each other. - Richard