Update predresults.R
I didn't make any direct progress on the "different" condition. However, I did make the filtering of groups and time variables their own functions, making the operation consistent. In doing so, the dataset is filtered into two dfs. This means there's one section of code to operate the mean/median and the separate dfs mean that it'll be simple to run the wilcox.test function on them later in the crate_results function. - Richard