Update predresults.R
Rewrote the create_results function, so that it's possible to use it as it's own independent function later, with error handling. In addition it's written, so that we can add the wilcox and normal functions for the 'difference' condition. Should easily wrap it up by the end of the week. - Richard