Update predresults.R
I've added t and wilcoxon tests as functions to be applied by create_results_vector. I took the code you had in the bottom of what you sent me and made this line:  accuracy[i] <- ifelse(p_val < phi_0, 1, 0), where accuracy is an empty vector being filled with 0s and 1s based on given predictions and p_val is the returned p value of the chosen test on the ith variable. Error handling needs some revision with so many arguments and conditions. Should be easier to do once the functions are in a separate file.