Update predresults.R
Cleaner, more modular and updated after we discussed the goals of the 'different' argument. Note that there is only a hypothesis argument, which can either be a string or a vector of strings with the valid hypotheses of 'different', 'increase' or 'decrease'. The error handling for that can be seen in custom_functions.R under get_results_vector