Update custom_functions.R
updated the function used to get the results vector, which is now named get_results_vector. It takes either a string or a vector of strings for the hypothesis argument, which must contain valid arguments of 'increase', 'decrease' or 'different'. returns 1s if correctly predicted and 0s if incorrectly predicted in the same index position as the variables